refactored setters in BBIntrin Doxygen tags for msb and lsb

+++ bitscan/bbtypes.h
@@ -24,7 +24,7 @@
 		
 using U8		= unsigned char;
 using U16		= unsigned short;
-using U32		= unsigned int;
+using U32		= unsigned long;
 using BITBOARD	= unsigned long long;
 using BYTE		= unsigned char;	
 using BOOL		= int;								//Deprecated: use bool instead (CHECK 08/02/25)
